requirements.txt
selenium==4.11.2
pandas==2.0.3
numpy==1.24.3
google-cloud-bigquery==3.11.3
python-dotenv==1.0.0
transformers==4.30.2
torch==2.0.1
